refactor(@dpc-sdp/nuxt-ripple): :recycle: refactor to call head functions from plugins
