+++ packages/nuxt-ripple/app.config.ts
@@ -33,6 +33,7 @@ export default defineAppConfig({
   ripple: {
     featureFlags: {},
     theme: {},
+    pageMiddleware: [],
     languages: {
       aii: {
         name: 'Noto Sans Syriac Eastern',
