+++ packages/nuxt-ripple/composables/use-tide-language.ts
@@ -1,5 +1,5 @@
 import { useAppConfig, useHead } from '#imports'
-import { computed, provide } from 'vue'
+import { computed } from 'vue'
 
 export default (page: any) => {
   const languages = useAppConfig()?.ripple?.languages
@@ -36,7 +36,11 @@ export default (page: any) => {
     })
   }
 
-  provide('language', { locale, direction, language })
+  useHead({
+    htmlAttrs: {
+      lang: language.value || 'en-AU'
+    }
+  })
 
   return { locale, direction, language }
 }
