+++ packages/nuxt-ripple/plugins/hide-alerts.ts
@@ -1,3 +1,5 @@
+import { useHead } from '#imports'
+
 /**
  * This script needs to be loaded as the very first asset on the page in order to avoid
  * the alerts showing that the user has already dismissed
@@ -44,4 +46,14 @@ try {
   console.error(e)
 }`
 
-export default hideAlertsOnLoadScript
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.hook('tide:page', () => {
+    useHead({
+      script: [
+        {
+          innerHTML: hideAlertsOnLoadScript
+        }
+      ]
+    })
+  })
+})
